Refactor utility functions into app.utils package
