+++ app/utils/__init__.py
@@ -1,0 +1,20 @@
+"""Utility functions for InvoiceManager."""
+
+from .activity import log_activity
+from .backup import create_backup, restore_backup
+from .imports import (
+    _import_csv,
+    _import_items,
+    _import_locations,
+    _import_products,
+)
+
+__all__ = [
+    'log_activity',
+    'create_backup',
+    'restore_backup',
+    '_import_csv',
+    '_import_items',
+    '_import_locations',
+    '_import_products',
+]
